BAU: Clarifying integration requirements for connecting services
There was some confusion over the requirements for Integration. Specifically,
an RP wanted to onboard/connect using an IP before having a domain available.
If skipping over the documentation it can be misleading as the last section on
the page could be interpreted as the only requirements for integration. This
commit adds some language to clarify this. Following our move to the *-a
environments in AWS and once HIDS monitoring is in place this requirement for
an IP address will be unnecessary.

solo: @jstandring-gds