Linked to VSP setup on GitHub
@RacingHippo suggested linking to information on setting up the VSP from the architecture page.
Docs on setting up the VSP are on GitHub, at the link I've put in.
This change brings the VSP in line with other components on the page which have links to documentation on setting them up. The only difference is that those components have the documentation on setting them up on the tech docs website, not GitHub.